+++ addon/-property-modifiers-mixin.js
@@ -1,7 +1,9 @@
 import { assert } from '@ember/debug';
+import { get } from '@ember/object';
 import Scheduler from './-scheduler';
 import {
   enqueueTasksPolicy,
+  enqueueWithPriorityPolicy,
   dropQueuedTasksPolicy,
   cancelOngoingTasksPolicy,
   dropButKeepLatestPolicy
@@ -22,6 +24,10 @@ export const propertyModifiers = {
 
   enqueue() {
     return setBufferPolicy(this, enqueueTasksPolicy);
+  },
+
+  enqueueWithPriority(sortFunc) {
+    return setBufferPolicy(this, Object.assign({ sortFunc }, enqueueWithPriorityPolicy));
   },
 
   drop() {
@@ -75,7 +81,7 @@ function assertModifiersNotMixedWithGroup(obj) {
 
 export function resolveScheduler(propertyObj, obj, TaskGroup) {
   if (propertyObj._taskGroupPath) {
-    let taskGroup = obj.get(propertyObj._taskGroupPath);
+    let taskGroup = get(obj, propertyObj._taskGroupPath);
     assert(`Expected path '${propertyObj._taskGroupPath}' to resolve to a TaskGroup object, but instead was ${taskGroup}`, taskGroup instanceof TaskGroup);
     return taskGroup._scheduler;
   } else {