+++ addon/-task-instance.js
@@ -5,6 +5,7 @@ import { run, join, schedule } from '@ember/runloop';
 import EmberObject, { computed, get, set } from '@ember/object';
 import Ember from 'ember';
 import {
+  cancelableSymbol,
   yieldableSymbol,
   YIELDABLE_CONTINUE,
   YIELDABLE_THROW,
@@ -209,11 +210,11 @@ let taskInstanceAttrs = {
   state: computed('isDropped', 'isCanceling', 'hasStarted', 'isFinished', function() {
     if (get(this, 'isDropped')) {
       return 'dropped';
-    } else if (get(this, 'isCanceling')) {
+    } else if (this.isCanceling) {
       return 'canceled';
-    } else if (get(this, 'isFinished')) {
+    } else if (this.isFinished) {
       return 'finished';
-    } else if (get(this, 'hasStarted')) {
+    } else if (this.hasStarted) {
       return 'running';
     } else {
       return 'waiting';
@@ -232,7 +233,7 @@ let taskInstanceAttrs = {
    * @readOnly
    */
   isDropped: computed('isCanceling', 'hasStarted', function() {
-    return get(this, 'isCanceling') && !get(this, 'hasStarted');
+    return this.isCanceling && !this.hasStarted;
   }),
 
   /**
@@ -346,10 +347,10 @@ let taskInstanceAttrs = {
    * @instance
    */
   cancel(cancelReason = ".cancel() was explicitly called") {
-    if (this.isCanceling || get(this, 'isFinished')) { return; }
+    if (this.isCanceling || this.isFinished) { return; }
     set(this, 'isCanceling', true);
 
-    let name = get(this, 'task._propertyName') || "<unknown>";
+    let name = (this.task && this.task._propertyName) || "<unknown>";
     set(this, 'cancelReason', `TaskInstance '${name}' was canceled because ${cancelReason}. For more information, see: http://ember-concurrency.com/docs/task-cancelation-help`);
 
     if (this.hasStarted) {
@@ -415,7 +416,8 @@ let taskInstanceAttrs = {
       value = new Error(this.cancelReason);
 
       if (this._debug || Ember.ENV.DEBUG_TASKS) {
-        Ember.Logger.log(this.cancelReason);
+        // eslint-disable-next-line no-console
+        console.log(this.cancelReason);
       }
 
       value.name = TASK_CANCELATION_NAME;
@@ -486,10 +488,10 @@ let taskInstanceAttrs = {
         this._triggerEvent('succeeded', this);
         break;
       case COMPLETION_ERROR:
-        this._triggerEvent('errored', this, get(this, 'error'));
+        this._triggerEvent('errored', this, this.error);
         break;
       case COMPLETION_CANCEL:
-        this._triggerEvent('canceled', this, get(this, 'cancelReason'));
+        this._triggerEvent('canceled', this, this.cancelReason);
         break;
     }
   },
@@ -544,7 +546,8 @@ let taskInstanceAttrs = {
     } finally {
       if (this._expectsLinkedYield) {
         if (!this._generatorValue || this._generatorValue._performType !== PERFORM_TYPE_LINKED) {
-          Ember.Logger.warn("You performed a .linked() task without immediately yielding/returning it. This is currently unsupported (but might be supported in future version of ember-concurrency).");
+          // eslint-disable-next-line no-console
+          console.warn("You performed a .linked() task without immediately yielding/returning it. This is currently unsupported (but might be supported in future version of ember-concurrency).");
         }
         this._expectsLinkedYield = false;
       }
@@ -707,7 +710,7 @@ let taskInstanceAttrs = {
       return;
     }
 
-    this._addDisposer(yieldedValue.__ec_cancel__);
+    this._addDisposer(yieldedValue[cancelableSymbol]);
 
     if (yieldedValue[yieldableSymbol]) {
       this._invokeYieldable(yieldedValue);
@@ -748,8 +751,8 @@ let taskInstanceAttrs = {
   _triggerEvent(eventType, ...args) {
     if (!this._hasEnabledEvents) { return; }
 
-    let host = get(this, 'task.context');
-    let eventNamespace = get(this, 'task._propertyName');
+    let host = this.task && this.task.context;
+    let eventNamespace = this.task && this.task._propertyName;
 
     if (host && host.trigger && eventNamespace) {
       host.trigger(`${eventNamespace}:${eventType}`, ...args);
@@ -775,15 +778,16 @@ taskInstanceAttrs[yieldableSymbol] = function handleYieldedTaskInstance(parentTa
   return function disposeYieldedTaskInstance() {
     if (yieldedTaskInstance._performType !== PERFORM_TYPE_UNLINKED) {
       if (yieldedTaskInstance._performType === PERFORM_TYPE_DEFAULT) {
-        let parentObj = get(parentTaskInstance, 'task.context');
-        let childObj = get(yieldedTaskInstance, 'task.context');
+        let parentObj = parentTaskInstance.task && parentTaskInstance.task.context;
+        let childObj = yieldedTaskInstance.task && yieldedTaskInstance.task.context;
         if (parentObj && childObj &&
             parentObj !== childObj &&
             parentObj.isDestroying &&
             get(yieldedTaskInstance, 'isRunning')) {
           let parentName = `\`${parentTaskInstance.task._propertyName}\``;
           let childName = `\`${yieldedTaskInstance.task._propertyName}\``;
-          Ember.Logger.warn(`ember-concurrency detected a potentially hazardous "self-cancel loop" between parent task ${parentName} and child task ${childName}. If you want child task ${childName} to be canceled when parent task ${parentName} is canceled, please change \`.perform()\` to \`.linked().perform()\`. If you want child task ${childName} to keep running after parent task ${parentName} is canceled, change it to \`.unlinked().perform()\``);
+          // eslint-disable-next-line no-console
+          console.warn(`ember-concurrency detected a potentially hazardous "self-cancel loop" between parent task ${parentName} and child task ${childName}. If you want child task ${childName} to be canceled when parent task ${parentName} is canceled, please change \`.perform()\` to \`.linked().perform()\`. If you want child task ${childName} to keep running after parent task ${parentName} is canceled, change it to \`.unlinked().perform()\``);
         }
       }
       yieldedTaskInstance.cancel();