+++ addon/-task-state-mixin.js
@@ -1,6 +1,6 @@
 import { gt } from '@ember/object/computed';
 import Mixin from '@ember/object/mixin';
-import { computed } from '@ember/object';
+import { computed, get } from '@ember/object';
 const { alias } = computed;
 
 // this is a mixin of properties/methods shared between Tasks and TaskGroups
@@ -40,15 +40,36 @@ export default Mixin.create({
   numQueued: 0,
   _seenIndex: 0,
 
-  cancelAll(reason = ".cancelAll() was explicitly called on the Task") {
+  cancelAll(options) {
+    let { reason, resetState } = options || {};
+    reason = reason || ".cancelAll() was explicitly called on the Task";
+
     this._scheduler.cancelAll(reason);
+
+    if (resetState) {
+      this._resetState();
+    }
   },
 
   group: computed(function() {
-    return this._taskGroupPath && this.context.get(this._taskGroupPath);
+    return this._taskGroupPath && get(this.context, this._taskGroupPath);
   }),
 
   _scheduler: null,
 
+  _resetState() {
+    this.setProperties({
+      'last': null,
+      'lastRunning': null,
+      'lastStarted': null,
+      'lastPerformed': null,
+      'lastSuccessful': null,
+      'lastComplete': null,
+      'lastErrored': null,
+      'lastCanceled': null,
+      'lastIncomplete': null,
+      'performCount': 0,
+    });
+  },
 });