+++ index.js
@@ -1,5 +1,5 @@
 'use strict';
 
 module.exports = {
-  name: 'ember-concurrency'
+  name: require('./package').name
 };